Remove re-catching of errors during wallet creation
The catch logic after a wallet creation is identical to what is defined anyway in the general Axios interceptor of the WaaS class.

This is noticeable because of https://github.com/Tangany/waas-js-sdk/issues/31, which required the activity id to be added to all the places that throw an error. This separate error-handling after wallet creation no longer makes sense.